+++ web/index_dev.php
@@ -4,5 +4,5 @@ require_once __DIR__.'/../vendor/autoload.php';
 
 Symfony\Component\Debug\Debug::enable();
 
-$app = new SKE\Application(__DIR__.'/../resources/config/dev.php');
+$app = new App\Application(__DIR__.'/../resources/config/dev.php');
 $app->run();
